requirements.txt
asgiref==3.10.0
channels==4.3.1
dj-database-url==3.0.1
Django==5.2.7
django-environ==0.12.0
django-pytest==0.2.0
djangorestframework==3.16.1
iniconfig==2.3.0
packaging==25.0
pluggy==1.6.0
psycopg==3.2.11
psycopg-binary==3.2.11
Pygments==2.19.2
pytest==8.4.2
pytest-django==4.11.1
python-dotenv==1.1.1
sqlparse==0.5.3
typing_extensions==4.15.0
